Uploaded Modification HessFrancis
Uploaded a version which terminates the function early when the function fails to provide a solution within 200 iterations. For accurate benchmarking, particularly for large matrices I would recommend increasing this number -preferably dynamically- as matrix size increases.
Please note the lines related to hessen functions should be altered when using this for Arnoldi.